sql migrate

+++ sql_models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, Integer, String, JSON, DateTime, ForeignKey
+from sqlalchemy import Boolean, Column, Integer, String, JSON, DateTime, ForeignKey, BigInteger
 from sqlalchemy.orm import relationship
 from database import Base
 import datetime
@@ -7,7 +7,7 @@ class User(Base):
     __tablename__ = "users"
 
     phone = Column(String, primary_key=True, index=True)
-    telegram_id = Column(Integer, unique=True, index=True, nullable=True)
+    telegram_id = Column(BigInteger, unique=True, index=True, nullable=True)
     tier = Column(String, default="free")
     trial_start_date = Column(DateTime, nullable=True)
     expiry_date = Column(DateTime, nullable=True)
@@ -30,7 +30,7 @@ class Feed(Base):
     user_id = Column(String, ForeignKey("users.phone"))
     name = Column(String)
     source_channel_ids = Column(JSON)  # List of integers
-    destination_channel_id = Column(Integer)
+    destination_channel_id = Column(BigInteger)
     active = Column(Boolean, default=True)
     delay_enabled = Column(Boolean, default=True)
     filters = Column(JSON, nullable=True)
